feat(password-service): refactor - moved packages to UserStoredKeysService and UserKeysService
