+++ mhpasswordmanager-password-service/src/main/java/br/com/marcoshssilva/mhpasswordmanager/passwordservice/domain/services/data/keys/exceptions/KeyRegistrationErrorException.java
@@ -1,4 +1,4 @@
-package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.data.keys;
+package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.data.keys.exceptions;
 
 public class KeyRegistrationErrorException extends Exception {
     public KeyRegistrationErrorException(String message, Throwable cause) {
